Rename compiler2 to compiler
