+++ pkg/compiler/analyze.go
